Fix karate on windows

+++ torch_geometric/datasets/karate.py
@@ -23,8 +23,8 @@ class KarateClub(InMemoryDataset):
         G = nx.karate_club_graph()
 
         adj = nx.to_scipy_sparse_matrix(G).tocoo()
-        row = torch.from_numpy(adj.row).to(torch.long)
-        col = torch.from_numpy(adj.col).to(torch.long)
+        row = torch.from_numpy(adj.row.astype(np.int64)).to(torch.long)
+        col = torch.from_numpy(adj.col.astype(np.int64)).to(torch.long)
         edge_index = torch.stack([row, col], dim=0)
         data = Data(edge_index=edge_index)
         data.num_nodes = edge_index.max().item() + 1
